created binary tree to circular doubly liked list
The challenge is to take an ordered binary tree and rearrange the internal pointers to make a circular doubly linked list out of it. The "small" pointer should play the role of "previous" and the "large" pointer should play the role of "next". The list should be arranged so that the nodes are in increasing order
http://cslibrary.stanford.edu/109/TreeListRecursion.html